Add Grid and Line classes; update Box class
Introduced a new `Grid` class with a constructor for creating grids and a `setColor` method for color customization. Modified the `Box` class to include a constructor and updated its `setColor` method. Added a new `Line` class that supports both rounded and rectangular lines, along with a method to set the origin to the center. Updated the `getArrow` function and included `constants.h` for constant values used in the new classes.